fix(ingest/browsePathsV2): Emit Container aspect first, to avoid BrowsePathsV2 generation race condition (#11813)
