+++ metadata-ingestion/src/datahub/ingestion/source/bigquery_v2/bigquery_schema_gen.py
@@ -523,7 +523,6 @@ class BigQuerySchemaGenerator:
                     logger.warning(
                         f"Could not create table ref for {table_item.path}: {e}"
                     )
-            yield from []
             return
 
         if self.config.include_tables:
